+++ bindings_node/src/client.rs
@@ -13,6 +13,7 @@ use tracing_subscriber::{fmt, prelude::*};
 pub use xmtp_api_grpc::grpc_api_helper::Client as TonicApiClient;
 use xmtp_cryptography::signature::ed25519_public_key_to_address;
 use xmtp_id::associations::builder::SignatureRequest;
+use xmtp_id::associations::verify_signed_with_public_context;
 use xmtp_mls::builder::ClientBuilder;
 use xmtp_mls::groups::scoped_client::LocalScopedGroupClient;
 use xmtp_mls::identity::IdentityStrategy;
@@ -43,7 +44,7 @@ impl Client {
 #[napi(string_enum)]
 #[derive(Debug)]
 #[allow(non_camel_case_types)]
-pub enum Level {
+pub enum LogLevel {
   off,
   error,
   warn,
@@ -52,9 +53,9 @@ pub enum Level {
   trace,
 }
 
-impl std::fmt::Display for Level {
+impl std::fmt::Display for LogLevel {
   fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-    use Level::*;
+    use LogLevel::*;
     let s = match self {
       off => "off",
       error => "error",
@@ -75,7 +76,7 @@ pub struct LogOptions {
   /// an option so that it does not require being specified in js object.
   pub structured: Option<bool>,
   /// Filter logs by level
-  pub level: Option<Level>,
+  pub level: Option<LogLevel>,
 }
 
 fn init_logging(options: LogOptions) -> Result<()> {
@@ -291,13 +292,46 @@ impl Client {
   }
 
   #[napi]
-  pub fn sign_with_installation_key(&self, text: String) -> Result<Vec<u8>> {
+  pub fn sign_with_installation_key(&self, signature_text: String) -> Result<Uint8Array> {
     let result = self
       .inner_client
       .context()
-      .public_sign(text)
+      .sign_with_public_context(signature_text)
       .map_err(ErrorWrapper::from)?;
 
-    Ok(result)
+    Ok(result.into())
+  }
+
+  #[napi]
+  pub fn verify_signed_with_installation_key(
+    &self,
+    signature_text: String,
+    signature_bytes: Uint8Array,
+  ) -> Result<()> {
+    let public_key = self.inner_client().installation_public_key();
+    self.verify_signed_with_public_key(signature_text, signature_bytes, public_key.into())
+  }
+
+  #[napi]
+  pub fn verify_signed_with_public_key(
+    &self,
+    signature_text: String,
+    signature_bytes: Uint8Array,
+    public_key: Uint8Array,
+  ) -> Result<()> {
+    let signature_bytes = signature_bytes.deref().to_vec();
+    let signature_bytes: [u8; 64] = signature_bytes
+      .try_into()
+      .map_err(|_| Error::from_reason("signature_bytes is not 64 bytes long."))?;
+
+    let public_key = public_key.deref().to_vec();
+    let public_key: [u8; 32] = public_key
+      .try_into()
+      .map_err(|_| Error::from_reason("public_key is not 32 bytes long."))?;
+
+    Ok(
+      verify_signed_with_public_context(signature_text, &signature_bytes, &public_key)
+        .map_err(ErrorWrapper::from)?,
+    )
   }
 }